refactor(components/molecule/accordion): remove defaultProps
