+++ components/molecule/accordion/demo/Demo.js
@@ -1,0 +1,11 @@
+import PropTypes from 'prop-types'
+
+const Demo = ({children}) => {
+  return <div style={{width: '100%', padding: 20}}>{children}</div>
+}
+
+Demo.propTypes = {
+  children: PropTypes.node
+}
+
+export default Demo
