+++ components/molecule/accordion/demo/DemoWrapper.js
@@ -1,0 +1,11 @@
+import PropTypes from 'prop-types'
+
+const DemoWrapper = ({children}) => {
+  return <div style={{display: 'flex', flexWrap: 'wrap'}}>{children}</div>
+}
+
+DemoWrapper.propTypes = {
+  children: PropTypes.node
+}
+
+export default DemoWrapper
