+++ components/molecule/accordion/demo/index.js
@@ -1,52 +1,13 @@
 import {useState} from 'react'
-import PropTypes from 'prop-types'
+
+import {text, icon, customTitle} from './settings.js'
 
 import MoleculeAccordion from '../src/index.js'
-
-const Demo = ({children}) => {
-  return <div style={{width: '100%', padding: 20}}>{children}</div>
-}
-
-Demo.propTypes = {
-  children: PropTypes.node
-}
-
-const DemoWrapper = ({children}) => {
-  return <div style={{display: 'flex', flexWrap: 'wrap'}}>{children}</div>
-}
-
-DemoWrapper.propTypes = {
-  children: PropTypes.node
-}
+import Demo from './Demo.js'
+import DemoWrapper from './DemoWrapper.js'
 
 export default () => {
   const [openedTabs, setOpenedTabs] = useState([1])
-  const text =
-    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam dictum magna non diam euismod blandit et eu ex. Vivamus pulvinar sodales tincidunt. Proin venenatis tristique quam, quis vehicula eros volutpat sed. Etiam sed tristique ante. Aenean commodo erat quis pulvinar luctus. Pellentesque ultricies lorem vitae ante euismod, at imperdiet est euismod. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. In dignissim porttitor sem, a varius nisl ullamcorper ut. Vivamus lacinia, quam eu placerat tempus, velit massa vulputate turpis, sit amet bibendum risus massa sit amet urna. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Praesent finibus lobortis blandit. Vivamus scelerisque blandit purus a suscipit. Nunc mi elit, condimentum eget pulvinar eu, lacinia vitae ligula. Sed sit amet eros auctor ipsum tincidunt hendrerit ac mollis justo. Ut ac sagittis ipsum.'
-
-  const icon = (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      x="0px"
-      y="0px"
-      width="16px"
-      height="16px"
-      viewBox="0 0 32 32"
-    >
-      <path d="M30,9l1,1c1,1,1,2,0,2L17,25c-1,1-2,1-2,0 L2,12c-1-1-1-2,0-2l1-1c1-1,2-1,2,0L16,20L28,9 C28,8,29,8,30,9z" />
-    </svg>
-  )
-  const customTitle = (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 8 8"
-    >
-      <path d="M4 0c-2.21 0-4 1.79-4 4s1.79 4 4 4 4-1.79 4-4-1.79-4-4-4zm-1.5 1.78l1.5 1.5 1.5-1.5.72.72-1.5 1.5 1.5 1.5-.72.72-1.5-1.5-1.5 1.5-.72-.72 1.5-1.5-1.5-1.5.72-.72z" />
-    </svg>
-  )
-
   return (
     <div>
       <h1>Accordion</h1>
