+++ components/molecule/accordion/demo/index.scss
@@ -1,0 +1,2 @@
+@import '~@s-ui/theme/lib/index';
+@import '~@s-ui/react-molecule-accordion/lib/index';
